Move PathExt::device_id() outside modules module
